+++ R/pat_internalFit.R
@@ -7,12 +7,12 @@
 #' @title Linear model fitting of channel A and B data
 #' 
 #' @param pat Purple Air Timeseries "pat" object
-#' @param subset (From `stats::lm`) An optional vector specifying a subset of
-#'   observations to be used in the fitting process.
-#' @param weights (From `stats::lm`) An optional vector of weights to be used in
-#'   the fitting process. Should be NULL or a numeric vector. If non-NULL,
-#'   weighted least squares is used with weights weights (that is, minimizing
-#'   sum(w*e^2)); otherwise ordinary least squares is used.
+# @param subset (From `stats::lm`) An optional vector specifying a subset of
+#   observations to be used in the fitting process.
+# @param weights (From `stats::lm`) An optional vector of weights to be used in
+#   the fitting process. Should be NULL or a numeric vector. If non-NULL,
+#   weighted least squares is used with weights weights (that is, minimizing
+#   sum(w*e^2)); otherwise ordinary least squares is used.
 #' @param showPlot logical specifying whether to generate a model fit plot
 #' @param size size of points
 #' @param shape symbol to use for points
@@ -32,8 +32,8 @@
 
 pat_internalFit <- function(
   pat = NULL,
-  subset = NULL,
-  weights = NULL,
+#  subset = NULL,
+#  weights = NULL,
   showPlot = TRUE,
   size = 1,
   shape = 15,
@@ -41,6 +41,11 @@ pat_internalFit <- function(
   alpha = 0.25,
   xylim = NULL
 ) {
+  
+  # TODO:  Decide to remove subset and weights and delete the code.
+  # Disable these parameters as too far down in the weeds.
+  subset <- NULL
+  weights <- NULL
   
   # ----- Validate parameters --------------------------------------------------