+++ R/pat_sample.R
@@ -42,7 +42,7 @@
 #' 
 #' @examples 
 #' \dontrun{
-#' pat <- pat_load()
+#' pat <- pat_loadLatest()
 #' subset <- pat_sample(pat, sampleSize=1000, setSeed=1234)
 #' }
 #'