+++ local_examples/03_pat_functions.R
@@ -17,9 +17,9 @@ pas_leaflet(scaqmd)
 
 # Let's get time series data for a sensor in Nipomo
 
-pat <- pat_load(pas, "SCNP_14",
-                startdate = "2019-03-01",
-                enddate = "2019-04-15")
+pat <- pat_loadLatest(pas, "SCNP_14",
+                      startdate = "2019-03-01",
+                      enddate = "2019-04-15")
 
 # Quick look at raw data to check for spurious correlations
 pat_scatterplot(pat)